+++ components/ColorPalette.ts
